Add and use function to handle input from user
Before, the game would exit abnormally if the user entered an
input that was not valid, like an empty string, a non-numeric input,
or an input out of range.

Now, the game is robust enough to keep asking the user for a new
entry until a valid one is provided.